refactor(client): extract useDownloadManager from P2PTransfer
Second step of decomposing P2PTransfer.tsx into focused hooks. Move the
receiver-side download orchestration (Download All + Download ZIP via fflate)
and its progress state (isZipping, isDownloading, downloadProgress) into
client/hooks/useDownloadManager.ts, along with the ReceivedFile type and the
fflate / dedupeFileName imports.

Behavior-preserving: the component passes receivedFiles (read-only) and
setError into the hook. The per-file <a download> links, the iOS ZIP tip, and
the receiver connection logic stay in the component.
